HTTP/2 Server Flow-control - window update strategies

+++ nima/http2/http2/src/main/java/io/helidon/nima/http2/WindowSize.java
@@ -78,9 +78,9 @@ public interface WindowSize {
     /**
      * Reset window size.
      *
-     * @param size window size
+     * @param size new window size
      */
-    void resetWindowSize(long size);
+    void resetWindowSize(int size);
 
     /**
      * Increment window size.
